requirements.txt
Flask
pandas
python-docx